Move PumiTallyImpl to a separate header
- Moved PumiTallyImpl class that encapsulates implementation details to a separate file.
- Updated test files to include the new PumiTallyImpl header.
- Adjusted CMakeLists.txt to link against the new implementation correctly.
- Disabled tests by default in CMake options.
